+++ src/extension/commands/run.ts
@@ -24,18 +24,47 @@ export async function importRequests() {
         // const pickType = await vscode.window.showQuickPick([importoptions.postman, importoptions.swagger2, importoptions.swagger3]) as importoptions;
         const pickType = importoptions.postman;
         if (!pickType) { return }
-        const link = await vscode.window.showInputBox({
-            prompt: "postman link",
-            ignoreFocusOut: true,
-            // validateInput: (value) => {
-            // if (value.startsWith("https://www.getpostman.com/collections") ||
-            //     value.startsWith("https://www.postman.com/collections")) {
-            //     return null;
-            // } else return "link should start with https://www.getpostman.com/collections/ or https://postman.com/collections";
-            // },
-            placeHolder: "https://getpostman.com/collections"
-        });
-        if (!link) { return }
+        const linkOrFile = await vscode.window.showQuickPick([{
+            label: "link",
+            description: "postman collection link",
+            picked: true,
+            alwaysShow: true,
+        }, {
+            label: "file",
+            description: "file exists in local system",
+            picked: false,
+            alwaysShow: true,
+        }]);
+        var filenameToimport: string | undefined;
+        if (linkOrFile) {
+            if (linkOrFile['label'] === 'link') {
+                filenameToimport = await vscode.window.showInputBox({
+                    prompt: "postman link",
+                    ignoreFocusOut: true,
+                    // validateInput: (value) => {
+                    // if (value.startsWith("https://www.getpostman.com/collections") ||
+                    //     value.startsWith("https://www.postman.com/collections")) {
+                    //     return null;
+                    // } else return "link should start with https://www.getpostman.com/collections/ or https://postman.com/collections";
+                    // },
+                    placeHolder: "https://getpostman.com/collections"
+                });
+            } else if (linkOrFile['label'] === 'file') {
+                const importUri = await vscode.window.showOpenDialog({
+                    canSelectFolders: false,
+                    canSelectFiles: true,
+                    title: "select file to import resource",
+                    filters: { "Postman Collection": ["json", "postman_collection.json"] },
+                    canSelectMany: false,
+                });
+                if (importUri && importUri.length > 0) {
+                    filenameToimport = importUri[0].fsPath;
+                }
+            }
+        } else { return }
+        if (!filenameToimport) {
+            return;
+        }
         const importUri = await vscode.window.showOpenDialog({
             canSelectFolders: true,
             canSelectFiles: false,
@@ -51,7 +80,7 @@ export async function importRequests() {
         await vscode.workspace.fs.createDirectory(folder);
         if (folder) {
             if (pickType === importoptions.postman) {
-                const result = await ApplicationServices.get().clientHanler.importPostman({ directory, link, save: true });
+                const result = await ApplicationServices.get().clientHanler.importPostman({ directory, link: filenameToimport!, save: true });
                 if (result.error == true) {
                     vscode.window.showErrorMessage(`import postman failed with error ${result.error_message}. 
 this usually happens for postman schema 1.0.0,